12.21. Desafio: Implementando o endpoint de emissão de pedidos

+++ src/main/resources/db/migration/V20250824112109__V004_cria-tabelas-pedido-item-pedido.sql
@@ -30,7 +30,7 @@ create table pedido (
 
 create table item_pedido (
                              id bigint not null generated always as identity,
-                             quantidade smallint not null,
+                             quantidade integer not null,
                              preco_unitario decimal(10,2) not null,
                              preco_total decimal(10,2) not null,
                              observacao varchar(255) null,
